correct linux versions
correct linux versions: we build mro packages from source, so all
the same versions of linux that work with other continuum packages
should work here. the MKL packages on the site are all the same
except for the one for SLES 10, which we don't support, so it isn't
actually important that they download the correct one for their
version of Linux, so removed that note.