Mark dot as allowed character in package name
Conda does not complain if dots are used as a part of the package name, yet this is not mentioned in the documentation.

Packages that have dots in name are already present in default channel (zope.deprecation, zope.interface, ...) as well as on conda-forge (bob.ap, bob.blit, bob.core, ...). There are also many python packages on pypi that include dot in the name that might be used in conda recipes in the future and would benefit if the same name could be used for both packages.