Make use of the CONDA_PREFIX environment variable
Note: some change would have to be made also for the Windows instructions. I don't know what that change would be, having never used the shell on Windows.

Using the CONDA_PREFIX environment would make it easier to create build scripts which require editing of variables.

 This seems to be alluded to elsewhere in the conda documentation:

https://conda.io/docs/user-guide/tasks/build-packages/environment-variables.html

Here is a real-world use case where I could imagine this would be helpful for creating an easy-to-use conda package:

https://github.com/n-riesco/ijavascript/blob/master/doc/install.md#conda

In particular, editing environment variables seems to be necessary whenever one wants to install both a conda package with NPM package dependencies (and have the NPM package be specific to the conda environment). This is a big deal e.g. for JupyterLab extensions. For example: https://github.com/jupyterlab/jupyterlab-latex#installation

In general, this slight change to the docs might make it easier for conda package creators to use NPM package dependencies, possibly better integrating NPM and conda and thereby increasing adoption of conda.