Update link-scripts.rst to be more explicit
Detailed exactly where the scripts need to be located and how they should be named. This makes it much clearer how to implement them, as the documentation has been implying that they should be within the source passed to conda-build and named appropriately. The example was also removed because simply placing bin/.<name>-<action>.sh into the .tar.bz2 archive does not lead to it getting executed.