miniconda.rst: Remove new section
A commit a month ago added a "Latest Miniconda Installer Links" section that's not built from the automation scripts around hashes. It also had a typo ("MaxOSX" rather than "MacOSX"). Earlier I did a pull request that just fixed the name, but looking at the automation around this, it's better to just remove the new section and leave users to go to their OS section like they used to.