Fix wget typo for installing miniconda on Linux
Encountered errors attempting to run the script downloaded as part of the instructions for installing miniconda on Linux. The `miniconda.sh` script file contained the wget logs which indicated an incorrect output option.

To rename a downloaded file with wget the option is `-O` not `-o` which outputs the log messages to the file.